requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
ldap3
sqlmodel
psycopg2-binary
pytest
